feat(controls): fontSize is a float on every text-bearing control
<Text> / <Btn> / <Tab> / <InputField> declared fontSize as int while
<Collapsible> / <TabMenu> / <Markdown> already took a float, so a
fontSize="6.5" on the former four hit int.Parse and killed the whole
Screen with a ParseException. TMP_Text.fontSize and
TMP_InputField.pointSize are floats — there was no engine reason for
the restriction.

- the four [UIAttr("fontSize")] setters become float
- MarkdigRenderer no longer casts BodySize through int, so
  <Markdown fontSize="16.5"> renders at 16.5 instead of a silent 16
- XSD: Text's hand-written fontSize is xs:float; Tab / InputField follow
  by reflection. New guard test: no control declares fontSize as xs:int
- XML skill rows + master spec note the float type

+++ Runtime/MarkdigBackend/MarkdigRenderer.cs
@@ -160,7 +160,7 @@ namespace PromptUGUI.MarkdigBackend
         {
             var n = new ElementNode("Text");
             n.Attributes["width"] = "stretch";
-            n.Attributes["fontSize"] = ((int)size).ToString(CultureInfo.InvariantCulture);
+            n.Attributes["fontSize"] = size.ToString(CultureInfo.InvariantCulture);
             n.Attributes["font"] = _style.BodyFont;
             n.Attributes["wrap"] = _style.ParagraphWrap ? "true" : "false";
             n.Attributes["align"] = "top-left";
